Move wav loader to asset module
