+++ loom/common/assets/assetsSound.cpp
@@ -27,7 +27,7 @@
 
 #include "stb_vorbis.h"
 #include "minimp3.h"
-#include "loom/engine/sound/wavloader.h"
+#include "wavloader.h"
 
 #ifdef _MSC_VER
 #define stricmp    _stricmp
@@ -38,7 +38,7 @@
 #endif
 
 extern "C" loom_allocator_t *gAssetAllocator;
-static loom_logGroup_t gSoundAssetGroup = { "soundAsset", 1 };
+loom_logGroup_t gSoundAssetGroup = { "soundAsset", 1 };
 
 void loom_asset_registerSoundAsset()
 {
