+++ loom/common/assets/wavloader.cpp
@@ -3,7 +3,7 @@
 #include "loom/common/core/log.h"
 #include "loom/common/utils/utEndian.h"
 
-extern loom_logGroup_t gLoomSoundLogGroup;
+extern loom_logGroup_t gSoundAssetGroup;
 
 typedef struct
 {
@@ -34,7 +34,7 @@ bool load_wav(const uint8_t* inData,
 {
     if (inData == NULL)
     {
-        lmLogError(gLoomSoundLogGroup, "No input data passed to wav loader");
+        lmLogError(gSoundAssetGroup, "No input data passed to wav loader");
         return false;
     }
     
@@ -46,7 +46,7 @@ bool load_wav(const uint8_t* inData,
         riff->header.chunkId[2] != 'F' ||
         riff->header.chunkId[3] != 'F')
     {
-        lmLogError(gLoomSoundLogGroup, "Bad wav file format");
+        lmLogError(gSoundAssetGroup, "Bad wav file format");
         return false;
     }
     
@@ -55,7 +55,7 @@ bool load_wav(const uint8_t* inData,
         riff->riffType[2] != 'V' ||
         riff->riffType[3] != 'E')
     {
-        lmLogError(gLoomSoundLogGroup, "Bad wav file format");
+        lmLogError(gSoundAssetGroup, "Bad wav file format");
         return false;
     }
     
@@ -63,7 +63,7 @@ bool load_wav(const uint8_t* inData,
     
     if (inDataLen < sizeof(chunk_header) + riff->header.chunkDataSize)
     {
-        lmLogError(gLoomSoundLogGroup, "Not enough data in wav buffer");
+        lmLogError(gSoundAssetGroup, "Not enough data in wav buffer");
         return false;
     }
     
@@ -84,7 +84,7 @@ bool load_wav(const uint8_t* inData,
             if (fmt->compressionType != 0x01)
             {
                 // This loader only supports simple PCM (8-bit or 16-bit, which should be the most common)
-                lmLogError(gLoomSoundLogGroup, "Unsupported wav format: 0x%2x. Recommend 8 or 16-bit PCM",
+                lmLogError(gSoundAssetGroup, "Unsupported wav format: 0x%2x. Recommend 8 or 16-bit PCM",
                            fmt->compressionType);
                 return false;
             }
